overlays: return the list of overlays instead of a set
The overlays was returning a set that was setting up nixpkgs.overlays
directly to a list. Doing it this way makes it impossible to use the
overlays from outside the system. This commit changes the overlays to
return the list of overlays and set up nixpkgs.overlays by importing the
overlays folder.