nixos: dynamically create users; generate their HM config through a function call (#52)
This commit moves away from setting the `home-manager.users.<name>` directly from the `hosts/<host>/home.nix` into setting it on a new NixOS option. The new option is set to a function, given the user and nixosConfig, returns a new function to be set directly on `home-manager.users.<name>`. This way, I can dynamically create users as I need and configure the home of these users using the same function.