nixos: fix the nixos-version broken by the nixpkgs pinning
When the nixpkgs is pinned using `builtins.fetchGit` instead of a
channel, nixos lose track of the commit because the fetcher removes the
`.git` folder. This changeset constructs the nixos-version back using
the `system.nixos.release` stiched with the first seven characters of
the revision in the JSON file describing the nixpkgs source.

closes #87