hosts/cratos: use the xps-9380 nixos hardware profile (#140)
I was getting black screens and random lockups when using the XPS 9370 profile. It turned out to be due to the changes specific to Kaby Lake that are being imported. This CPU is actually Whiskey Lake. The profile also disables a bug fix that is not present on this model.

See https://github.com/NixOS/nixos-hardware/pull/97

fixes #138

This reverts commit 6f1d472a0cbdaa1bc4d182ec030865d534e42a0e