hosts/hera: update the configuration of autorandr (#164)
I finally got my two screen setup at work. This change updates the configuration of autorandr to work automatically on my desk.

The change has also fixed an issue with the post-switch script, where the primary screen was hardcoded as opposed to using the primary field from the xrandr output.