[GnuPG] Enable extra support (#210)
I use a YubiKey for all my GnuPG, and SSH needs. This change enables an extra GnuPG socket to be forwarded by SSH to a remote host. The GnuPG on the remote will then be able to use that socket to sign and encrypt. Below, is an example of a host configuration to forward the SSH agent between two NixOS hosts, same user, and the same uid.

```conf
Host myserver
  HostName myserver.example.com
  ExitOnForwardFailure yes
  RemoteForward /run/user/2000/gnupg/S.gpg-agent /run/user/2000/gnupg/S.gpg-agent.extra
```

I have not been able to resolve the lack of SSH agent on the remote host, however, because the local agent's socket for SSH is mounted at a different location than the one I'm currently sending over there. I still have to use the `-A` flag, but it's not a big deal.

It's important to take note of one limitation: This will not work for sending the socket to a host having the GnuPG running though.