nixos/darwin: switch to using binary cache over HTTP (#313)
* nixos: use binary caches over HTTP

By using caches over HTTP, it allows us to spawn up a local binary cache
for bandwidth savings. Although HTTP does not prevent MITM attacks, all
of the cached binaries are signed. If the signature does not match, nix
will reject it and move on to the next cache.

* darwin: use binary caches over HTTP

By using caches over HTTP, it allows us to spawn up a local binary cache
for bandwidth savings. Although HTTP does not prevent MITM attacks, all
of the cached binaries are signed. If the signature does not match, nix
will reject it and move on to the next cache.

* do not include cache.nixos.org public key as it is already included