modules: home: workstation: i3: polybar: disable warning log output
This is a temporary fix. A more permanent one would be to choose a font
supporting all characters used by our polybar configuration. But for
now, this is required so we don't have a bloated journal due to polybar
warnings about unmatched characters.

Adding TODO in #247 to improve the font used.