zsh: gcim: disable commit signature on remote systems
This commit disables the signing of Git commits on remote systems, by
checking the value of $SSH_AUTH_SOCK.

As a side effect, this also disables commit signing on remote systems
with GnuPG socket forwarded. As a todo item, we should somehow check for
this edge case and leave it enabled.